http: on error api - single action per code
With this patch, users can add only one action for
code. In a later patch, we will allow to branch further.

In implementation, we have a hashtable code->page, where page will
describes a further matching based on match string. For now only one
matchstring is stored in a page.

Signed-off-by: Antal Nemes <antal.nemes@balabit.com>